feat: Updates Makefile to consider a different execution of tests for each platform and the internal tests, refactors codeaccordingly. Fixes golangci-lint issues. Alsofixes issue were a more updated version of terraform, resulted in failing test as it was executed way faster during the time now comparison in our validation to check if certificate to be renewed. Adding one more day to make sure we always need renewal. Makes VaaS and TPP tests parallel
VC-38188
