+++ venafi/resource_venafi_policy_tpp_test.go
@@ -1,3 +1,6 @@
+//go:build tpp
+// +build tpp
+
 package venafi
 
 import (
@@ -24,9 +27,6 @@ variable "TPP_URL" {default = "%s"}
 variable "TPP_ZONE" {default = "%s"}
 variable "TPP_ZONE_ECDSA" {default = "%s"}
 variable "TRUST_BUNDLE" {default = "%s"}
-variable "CLOUD_URL" {default = "%s"}
-variable "CLOUD_APIKEY" {default = "%s"}
-variable "CLOUD_ZONE" {default = "%s"}
 variable "TPP_ACCESS_TOKEN" {default = "%s"}
 `,
 		os.Getenv("TPP_USER"),
@@ -35,9 +35,6 @@ variable "TPP_ACCESS_TOKEN" {default = "%s"}
 		os.Getenv("TPP_ZONE"),
 		os.Getenv("TPP_ZONE_ECDSA"),
 		os.Getenv("TRUST_BUNDLE"),
-		os.Getenv("CLOUD_URL"),
-		os.Getenv("CLOUD_APIKEY"),
-		os.Getenv("CLOUD_ZONE"),
 		os.Getenv("TPP_ACCESS_TOKEN"))
 
 	tokenProv = envVariables + `
@@ -47,25 +44,6 @@ provider "venafi" {
 	zone = "${var.TPP_ZONE}"
 	trust_bundle = "${file(var.TRUST_BUNDLE)}"
 }`
-
-	vaasProv = envVariables + `
-provider "venafi" {
-	url = "${var.CLOUD_URL}"
-	api_key = "${var.CLOUD_APIKEY}"
-}
-`
-
-	vaasPolicyResourceTest = `
-%s
-resource "venafi_policy" "vaas_policy" {
-	provider = "venafi"
-	zone="%s"
-	policy_specification = file("%s")
-}
-output "policy_specification" {
-	value = "${venafi_policy.vaas_policy.policy_specification}"
-}`
-
 	tppPolicyResourceTest = `
 %s
 resource "venafi_policy" "tpp_policy" {
@@ -76,7 +54,6 @@ resource "venafi_policy" "tpp_policy" {
 output "policy_specification" {
 	value = "${venafi_policy.tpp_policy.policy_specification}"
 }`
-
 	readPolicy = `
 %s
 resource "venafi_policy" "read_policy" {
@@ -86,215 +63,8 @@ resource "venafi_policy" "read_policy" {
 }`
 )
 
-//-----------------------------------------------VaaS test cases begins----------------------------------------------//
-
-func TestCreateVaasEmptyPolicy(t *testing.T) {
-	data := testData{}
-	data.zone = RandAppName() + "\\\\" + RandCitName()
-
-	data.filePath = GetAbsoluteFIlePath(emptyPolicy)
-
-	config := fmt.Sprintf(vaasPolicyResourceTest, vaasProv, data.zone, data.filePath)
-	t.Logf("Testing Creating empty Zone:\n %s", config)
-	resource.Test(t, resource.TestCase{
-		ProviderFactories: testAccProviderFactories,
-		Steps: []resource.TestStep{
-			{
-				Config: config,
-				Check: func(s *terraform.State) error {
-					t.Log("Creating VaaS empty zone: ", data.zone)
-					return checkCreateVaasPolicy(t, &data, s, false)
-				},
-			},
-		},
-	})
-}
-
-func TestCreateVaasPolicy(t *testing.T) {
-	data := testData{}
-	data.zone = RandAppName() + "\\\\" + RandCitName()
-
-	data.filePath = GetAbsoluteFIlePath(policySpecVaas)
-
-	config := fmt.Sprintf(vaasPolicyResourceTest, vaasProv, data.zone, data.filePath)
-	t.Logf("Testing creating VaaS Zone:\n %s", config)
-	resource.Test(t, resource.TestCase{
-		ProviderFactories: testAccProviderFactories,
-		Steps: []resource.TestStep{
-			{
-				Config: config,
-				Check: func(s *terraform.State) error {
-					t.Log("Creating VaaS zone: ", data.zone)
-					return checkCreateVaasPolicy(t, &data, s, false)
-				},
-			},
-		},
-	})
-}
-
-func checkCreateVaasPolicy(t *testing.T, data *testData, s *terraform.State, validateAttr bool) error {
-	t.Log("Validate Creating VaaS empty policy", data.zone)
-
-	pstUntyped := s.RootModule().Outputs["policy_specification"].Value
-
-	ps, ok := pstUntyped.(string)
-	if !ok {
-		return fmt.Errorf("output for \"policy_specification\" is not a string")
-	}
-
-	bytes := []byte(ps)
-
-	var policySpecification policy.PolicySpecification
-	err := json.Unmarshal(bytes, &policySpecification)
-	if err != nil {
-		return fmt.Errorf("policy specification is nil")
-	}
-
-	if !validateAttr {
-		return nil
-	}
-
-	//get policy on directory.
-	file, err := os.Open(data.filePath)
-	if err != nil {
-		return err
-	}
-
-	fileBytes, err := io.ReadAll(file)
-	if err != nil {
-		return err
-	}
-
-	var filePolicySpecification policy.PolicySpecification
-	err = json.Unmarshal(fileBytes, &filePolicySpecification)
-	if err != nil {
-		return err
-	}
-
-	equal := IsArrayStringEqual(filePolicySpecification.Policy.Domains, policySpecification.Policy.Domains)
-	if !equal {
-		return fmt.Errorf("domains are different, expected %+q but get %+q", filePolicySpecification.Policy.Domains, policySpecification.Policy.Domains)
-	}
-
-	//compare some attributes.
-
-	if *(filePolicySpecification.Policy.MaxValidDays) != *(policySpecification.Policy.MaxValidDays) {
-		return fmt.Errorf("max valid period is different, expected %s but get %s", strconv.Itoa(*(filePolicySpecification.Policy.MaxValidDays)), strconv.Itoa(*(policySpecification.Policy.MaxValidDays)))
-	}
-
-	equal = IsArrayStringEqual(filePolicySpecification.Policy.KeyPair.KeyTypes, policySpecification.Policy.KeyPair.KeyTypes)
-
-	if !equal {
-		return fmt.Errorf("key types are different, expected %+q but get %+q", filePolicySpecification.Policy.KeyPair.KeyTypes, policySpecification.Policy.KeyPair.KeyTypes)
-	}
-
-	equal = IsArrayStringEqual(filePolicySpecification.Policy.Subject.Countries, policySpecification.Policy.Subject.Countries)
-
-	if !equal {
-		return fmt.Errorf("countries are different, expected %+q but get %+q", filePolicySpecification.Policy.Subject.Countries, policySpecification.Policy.Subject.Countries)
-	}
-
-	if *(filePolicySpecification.Default.Subject.Locality) != *(policySpecification.Default.Subject.Locality) {
-		return fmt.Errorf("default locality is different, expected %s but get %s", *(filePolicySpecification.Default.Subject.Locality), *(policySpecification.Default.Subject.Locality))
-	}
-
-	return nil
-}
-
-func TestImportVaasPolicy(t *testing.T) {
-	config := getImportVaasConfig()
-	t.Logf("Testing importing VaaS Zone:\n %s", config)
-	resource.Test(t, resource.TestCase{
-		ProviderFactories: testAccProviderFactories,
-		Steps: []resource.TestStep{
-			{
-				Config:        config,
-				ResourceName:  "venafi_policy.read_policy",
-				ImportStateId: os.Getenv("CLOUD_POLICY_SAMPLE"),
-				ImportState:   true,
-				ImportStateCheck: func(states []*terraform.InstanceState) error {
-					t.Logf("Checking zone: %s's attributes", os.Getenv("CLOUD_POLICY_SAMPLE"))
-					return checkImportVaasPolicy(states)
-				},
-			},
-		},
-	})
-}
-
-func checkImportVaasPolicy(states []*terraform.InstanceState) error {
-	st := states[0]
-	attributes := st.Attributes
-
-	ps := attributes["policy_specification"]
-	bytes := []byte(ps)
-
-	var policySpecification policy.PolicySpecification
-	err := json.Unmarshal(bytes, &policySpecification)
-	if err != nil {
-		return fmt.Errorf("policy specification is nil")
-	}
-
-	//get policy on directory.
-	path := GetAbsoluteFIlePath(policySpecVaas)
-	file, err := os.Open(path)
-	if err != nil {
-		return err
-	}
-
-	fileBytes, err := io.ReadAll(file)
-	if err != nil {
-		return err
-	}
-
-	var filePolicySpecification policy.PolicySpecification
-	err = json.Unmarshal(fileBytes, &filePolicySpecification)
-	if err != nil {
-		return err
-	}
-
-	equal := IsArrayStringEqual(filePolicySpecification.Policy.Domains, policySpecification.Policy.Domains)
-	if !equal {
-		return fmt.Errorf("domains are different, expected %+q but get %+q", filePolicySpecification.Policy.Domains, policySpecification.Policy.Domains)
-	}
-
-	//compare some attributes.
-
-	if *(filePolicySpecification.Policy.MaxValidDays) != *(policySpecification.Policy.MaxValidDays) {
-		return fmt.Errorf("max valid period is different, expected %s but get %s", strconv.Itoa(*(filePolicySpecification.Policy.MaxValidDays)), strconv.Itoa(*(policySpecification.Policy.MaxValidDays)))
-	}
-
-	equal = IsArrayStringEqual(filePolicySpecification.Policy.KeyPair.KeyTypes, policySpecification.Policy.KeyPair.KeyTypes)
-
-	if !equal {
-		return fmt.Errorf("key types are different, expected %+q but get %+q", filePolicySpecification.Policy.KeyPair.KeyTypes, policySpecification.Policy.KeyPair.KeyTypes)
-	}
-
-	equal = IsArrayStringEqual(filePolicySpecification.Policy.Subject.Countries, policySpecification.Policy.Subject.Countries)
-
-	if !equal {
-		return fmt.Errorf("countries are different, expected %+q but get %+q", filePolicySpecification.Policy.Subject.Countries, policySpecification.Policy.Subject.Countries)
-	}
-
-	if *(filePolicySpecification.Default.Subject.Locality) != *(policySpecification.Default.Subject.Locality) {
-		return fmt.Errorf("default locality is different, expected %s but get %s", *(filePolicySpecification.Default.Subject.Locality), *(policySpecification.Default.Subject.Locality))
-	}
-
-	return nil
-}
-
-func getImportVaasConfig() string {
-	path := GetAbsoluteFIlePath(policySpecVaas)
-	zone := os.Getenv("CLOUD_POLICY_SAMPLE")
-	zone = strings.Replace(zone, "\\", "\\\\", 1)
-	config := fmt.Sprintf(readPolicy, vaasProv, zone, path)
-	return config
-}
-
-//------------------------------------------------VaaS test cases ends------------------------------------------------//
-
-//------------------------------------------------TPP test cases begins-----------------------------------------------//
-
-func TestCreateTppEmptyPolicy(t *testing.T) {
+func TestTPPCreateEmptyPolicy(t *testing.T) {
+	t.Parallel()
 	data := testData{}
 	rootZone := os.Getenv("TPP_PM_ROOT")
 	rootZone = strings.Replace(rootZone, "\\", "\\\\", 4)
@@ -318,7 +88,8 @@ func TestCreateTppEmptyPolicy(t *testing.T) {
 	})
 }
 
-func TestCreateTppPolicy(t *testing.T) {
+func TestTPPCreatePolicy(t *testing.T) {
+	t.Parallel()
 	data := testData{}
 	rootZone := os.Getenv("TPP_PM_ROOT")
 	rootZone = strings.Replace(rootZone, "\\", "\\\\", 4)
@@ -335,14 +106,15 @@ func TestCreateTppPolicy(t *testing.T) {
 				Config: config,
 				Check: func(s *terraform.State) error {
 					t.Log("Creating TPP zone: ", data.zone)
-					return checkCreateVaasPolicy(t, &data, s, false)
+					return checkCreatePolicy(t, &data, s, false)
 				},
 			},
 		},
 	})
 }
 
-func TestImportTppPolicy(t *testing.T) {
+func TestTPPImportPolicy(t *testing.T) {
+	t.Parallel()
 	config := getPolicyImportTppConfig()
 	t.Logf("Testing importing TPP Zone:\n %s", config)
 	resource.Test(t, resource.TestCase{
@@ -515,5 +287,3 @@ func checkImportTppPolicy(states []*terraform.InstanceState) error {
 
 	return nil
 }
-
-//-------------------------------------------------TPP test cases ends------------------------------------------------//
